requirements.txt
aiohttp
beautifulsoup4
python-telegram-bot==20.0
